Add LSH256 and LSH512 dynamic dispatch
